+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## v1.6
+
+Added
+
+- Saint Andral's Church
+- Coffin Maker's Shop
+
+Residents
+
+- Father Lucian Petrovich
+- Yeska
+- Milivoj
+- Henrik van der Voort
+- Unknown Vampire Spawn
+
+Assets
+
+- Saint Andral's Church
+- Father Lucian Petrovich
+- Yeska
+- Coffin Maker's Shop
+- Unknown Vampire Spawn
+
 ## v1.5
 
 Added